+++ docs/CHANGELOG.rst
@@ -13,6 +13,8 @@ Unreleased
 Added
 -----
 - Add GEOparse to the ``resolwebio/common:2.7.0`` Docker image
+- Add fastq file validation in ``import-sra-single`` and
+  ``import-sra-paired`` processes
 
 Changed
 -------